requirements.txt
hvac==0.10.14
pyyaml==5.3.1
termcolor==1.1.0